FIX: Don't crash pluginConsequence on bad CDS
If a CDS is not found in the gtf file, pluginConsequence crashes. This
may be ok for debugging purposes (as it may indicate a bug or bad data),
but will hold up an entire sVEP run on a single line.

If the CDS is not present in the gtf file, log a warning and fill in the
missing bases with some dummy bases, which will closely mimick the
likely consequence of the mutation, without having access to the
truncated bases.