FIX: Rebuild on any pluginConsequence change
The docker image is only rebuilt when a change in the
pluginConsequence/docker directory occurs. This means if other files
(such as any of the files in the consequence directory) are altered,
these changes are not synced until a docker/ change is also made.

Increase the scope of files tracked for the purposes of triggering a
rebuilding of the pluginConsequence docker image to the entire
directory.
TODO: Move this rebuilding into the lambda module itself, rather than as
part of the lambda directory.