FIX: Add support for end of CDS deletions
A previous commit 9ef004b0467f19b360e7b2a4a1d44e11061d61f3 Added support
for SNPs whose codon effects required adjacent exons to describe, as
they were located in a codon that stretched across two exons. This only
took into account the position of the SNP, so deletions could be missed
if they started in a codon fully contained within a CDS but extended
into a partial codon.

Add support for deletions by inspecting the end of the mutation (as
calculated by the length of the reference allele) and allow padding to be
added to both the start and the end of the reference if required (for
deletions that span the whole exon).