FIX: Mark incomplete alt codons as not inframe_del
The inframe_deletion function doesn't check to see if the replacement
for the the complete reference codons is itself a complete set of
codons.

Ensure that the difference between the reference and alt codons is
divisible by 3. If it isn't, then the mutation is not inframe and
shouldn't be marked as such.