FIX: Don't assume alt codon is complete.
When the alt codon is translated, the codon is marked as complete. If
the codon is only two base pairs long, this implies that there is a
third base that will complete the codon and the peptide will be
calculated from the available 2 bases if possible. This deviates from
VEP's behaviour, which marks any incomplete codon as unknown.

Don't mark incomplete alt codons as complete, to retain the same
behaviour as VEP.