FEAT: Use VEP output style for insertion positions
If there is an insertion between base A and base B, VEP marks the
position as A-B, despite requiring input to mark the mutation in
reversed notation B-A. sVEP currently uses the reversed notation in its
output.

Change the position output to match VEP's output, by ensuring the lower
base appears first in the position string.