FIX: Correct frameshift calculation
The frameshift calculation uses the seq_length value, which it evidently
expects to be the length of the CDS after the alt allele has been
applied. However this value actually holds the length of the CDS for the
ref, which is always the same value as the other value to which the
function compares. This means that nothing is considered a frameshift
mutation.

As the refCodon and altCodons have been filled in at this point, these
values can be used. If they haven't been set, call
_get_peptide_alleles() to set them. refCodon should always be a complete
set of codons, and altCodon is those codons after the alternate allele
is applied. Therefore we can uses the difference in these lengths to
calculated if a mutation is a frameshift mutation. To maintain the same
behaviour as the previous function, the remainder is returned.