FIX: Fail frameshift calculation gracefully
When calculating whether a feature is a frameshift, the refCodon and
altCodon feat attributes are required. These should be set by
_get_peptide_alleles(), but if _get_peptide_alleles() exits prematurely
due to an edgecase these are not set. This causes problems for the
downstream frameshift function.

If the values are not correctly defined, exit frameshift with a 0 value.
Note that the output will also likely come with a warning in this case
that should be addressed.