REFACTOR: Replace chrom_mapping with refChrom
Every lambda function downstream of initQuery is passed the mapping
dictionary, which contains the chromosomes in the VCF and what the name
of the chromosome is likely to be in the reference files. For example an
element might be "chr12": "12". However as only queryVCF actually deals
with more than one chromosome, most of the dictionary is useless. Even
if a plugin did need to look up another chromosome, it wouldn't know what
key to use as access to the VCF is lost after queryVCF. In addition, in
order to know what reference to access, the input first needs to be
interrogated to find which chromosome is relevant to the function.

Replace the chrom_mapping dictionary with the refChrom string, which is
just as effective, and avoids the need to interrogate the main data
before downloading required reference files.