FEAT: Add ref, gt, qual and filter columns
Additional columns have been requested for sVEP. These are the ref
allele, the genotype (e.g. 0/0, 0/1, 1/1), the VCF QUAL column, and the
VCF FILTER column.

Add the genotype as gt, the QUAL column as qual, and the FILTER column
as filter, pulled  directly from the VCF. Use the pluginConsequence's
internal $ref variable as the ref value, as it has been processed in the
same way as the alt. These are currently only included in the backend,
and lack frontend access.