FIX: Find indels with clinvar and gnomad plugins
The clinvar and gnomad plugins use the variant details from
pluginConsequence to match the bedfile and VCF respectively. While this
works for SNPs, indels are represented differently and are missed.

Use the VCF representations of pos, ref and alt for clinvar and gnomad
plugins. This entails separating multiallelic records inf queryVcf and
passing the VCF formatted information through the pipeline. In order to
simplify this process, also switch subprocess.Popen calls to use the
CheckedProcess wrapper. Note that posVcf is an int, rather than a
string, although changes in pluginConsequence might unexpectedly change
that behaviour. Also fix the logging and timeout behaviour for
pluginGnomad to handle each query separately, so lines that aren't in
the gnomad VCF aren't automatically passed to future invocations.