REFACTOR: Query each clinvar location once.
The clinvar bedfile is queried for every line, despite the fact that
many lines originate from the same location, with for example different
transcripts.

Group lines by position before querying the bedfile, to substantially
reduce the amount of querying. This could be further reduced by querying
multiple locations simultaneously, similarly to the pluginGnomad
function, but query numbers are less important here because the file is
local.