FIX: Create job before starting sVEP pipeline
initQuery starts the downstream functions before creating the job item
in dynamodb. This creates a race condition, however unlikely, that
another function may try to edit the job before it is created (for
example if throttling from other pipelines interferes with initQuery's
dynamoDB call).

Create the job item before starting any functions that may attempt to
interact with that item.