REFACTOR: Consolidate sVEP pipeline boilerplate
Most of the sVEP pipeline functions pass around the same variables;
tempFileName, requestId and refChrom. They also all wrap most of their
code in a try/except block with the exception updating the job in
dynamodb. Functions that call themselves when running out of time carry
around all their arguments in case they need to call themselves.

Move the try/except block into an orchestration context, that handles
marking functions complete as well as updating the dynamodb item on
failure. Add handling of the common fields to the Orchestration instance
itself. Automatically increment the filename on repeated calls and
assign pseudo filenames to functions that aren't tracked for
concatenation purposes. This helps with logging and possible s3 storage
of large payloads. Create a convenience function to resend a message to
the current lambda function in the case of imminent timeouts. This
function will use the initial message, with custom additions and a new
filename. This function also doesn't need to be passed the SNS topic
arn, as it was present in the initial event, so remove these from
environment variables.