REFACTOR: Handle next svep step in orchestrator
Individual sVEP pipeline functions don't need to know the identity of
the next function, other than the name of the next SNS topic. This is
passed in as "functionname_SNS_TOPIC_ARN".

Pass the next function's SNS topic arn in as instead
NEXT_FUNCTION_SNS_TOPIC_ARN. Then handle the lookup of this environment
variable in the Orchestrator object. As this is used for tracked
functions, change the default value of track for start_function to
False, so the argument can be removed from most other calls to
start_function. This means adding additional plugins does not require
changes to the function code of the surrounding functions, only the IAM
roles and NEXT_FUNCTION_SNS_TOPIC_ARN environment variable.