FIX: Only trigger reference update on change
The comparison that is used to trigger updating the reference files
compares the version or hash string to the full response from dynamodb.
These are always going to be different, so references are re-downloaded
every time.

Compare the version or hash string specifically against the version
attribute of the relevant item.