FEAT: Automatically update clinvar reference
The clinvar.bed.gz(.csi) reference is computed and uploaded manually,
which causes problems with new deployments and means it is often
outdated.

Hook the clinvar reference into the automatic reference updating of
updateReferenceFiles, so it is updated with each run of the function, as
long as the version has changed. Note that the pipeline to create the
reference would take over 30 minutes on the lambda function, so instead
of sending an SNS back to itself, make updateReferenceFiles run an EC2
instance to run the pipeline.