FIX: Only mark clinvar updated after upload
The clinvar reference update takes over half an hour, during which
something can go wrong and the reference not be updated. However the
dynamodb item is updated to the new version immediately. This means the
version may read that the reference has been updated, while the old
version's file is still in use.

Move the dynamodb version update for the clinvar reference to the end of
the EC2 run, after the new files have been uploaded. This change was not
made to the other two references, because the gnomad reference update is
fanned out, and the mirna reference is unused anyway.