FIX: Allow clearTempAndRegions to finish
When there are many files to delete, clearTempAndRegions can run out of
retries before deleting all the files. In addition, if a job has failed,
other parts of the job may continue to run, producing additional files.

Allow clearTempAndRegions to run as long as it needs, calling itself
again when out of time. Also call this function whenever a job moves
into a state that means the component files should be deleted (currently
completed, expired, failed), which means concatPages no longer needs to
explicitly call it.