FIX: Circumvent rare crashes in pluginConsequence
When features are the UTR of a gene, sometimes they can cause svep to
crash, as we haven't implemented a robust handling of these edge cases
yet. In addition, some empty lines are sneaking through and causing
crashes.

Circumevent the UTR calculation logic calls in the vep code, to prevent
the pipeline from crashing. These regions are typically filtered out
downstream anyway. Also assign a very high rank to empty lines to cause
them to be filtered out.