FIX: Give queryVcf more time to wrap up
When queryVcf has a large amount of input data, it tries to cut it up
into smaller chunks to process in parallel. If there is too much data,
even this chopping up and publishing to SNS can take to long, causing
the function to time out.

Allow more time for the function to send the remaining data in batches
to queryVcfSubmit.